test(app): update typography e2e test to work
