+++ src/components/app/test/typography/pages/page-one/page-one.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { IonicPage } from '../../../../../..';
+
+@IonicPage({
+  name: 'page-one'
+})
+@Component({
+  templateUrl: 'page-one.html'
+})
+export class PageOne { }
